Increased coverage to >90%
