+++ contracts/test/exchange/TestUniswapFactory.sol
@@ -1,0 +1,27 @@
+pragma solidity ^0.5.10;
+import "./TestUniswapExchange.sol";
+import "../../exchange/UniswapFactory.sol";
+
+
+contract TestUniswapFactory is UniswapFactory {
+
+    function testLaunchExchange(
+        address _token
+    ) public returns (address exchange) {
+        require(
+            tokenToExchange[_token] == address(0),
+            "Already an exchange for that."
+        ); //There can only be one exchange per token
+        require(
+            _token != address(0) && _token != address(this),
+            "Not a valid token address."
+        );
+        address payable newExchange = address(new TestUniswapExchange(_token));
+        tokenList.push(_token);
+        tokenToExchange[_token] = newExchange;
+        exchangeToToken[newExchange] = _token;
+        emit ExchangeLaunch(newExchange, _token);
+        return newExchange;
+    }
+
+}
