+++ test/exchange/UniswapExchange.01.test.ts
@@ -4,13 +4,13 @@ const { balance, BN, constants, ether, expectEvent, expectRevert, send } = requi
 import {
     TestERC20MintableInstance,
     TestUniswapExchangeInstance,
-    UniswapFactoryInstance,
+    TestUniswapFactoryInstance,
 } from '../../types/truffle-contracts';
 
 // tslint:disable:max-line-length
 const TestERC20Mintable = artifacts.require('./test/issuance/TestERC20Mintable.sol') as Truffle.Contract<TestERC20MintableInstance>;
 const UniswapExchange = artifacts.require('./test/exchange/TestUniswapExchange.sol') as Truffle.Contract<TestUniswapExchangeInstance>;
-const UniswapFactory = artifacts.require('./exchange/UniswapFactory.sol') as Truffle.Contract<UniswapFactoryInstance>;
+const UniswapFactory = artifacts.require('./test/exchange/TestUniswapFactory.sol') as Truffle.Contract<TestUniswapFactoryInstance>;
 // tslint:enable:max-line-length
 
 // tslint:disable-next-line:no-var-requires
@@ -21,7 +21,7 @@ chai.should();
 contract('UniswapExchange - Initialization', (accounts) => {
 
     let token: TestERC20MintableInstance;
-    let uniswapFactory: UniswapFactoryInstance;
+    let uniswapFactory: TestUniswapFactoryInstance;
     let uniswapExchange: TestUniswapExchangeInstance;
     const initialiser1 = accounts[1];
 
@@ -32,7 +32,7 @@ contract('UniswapExchange - Initialization', (accounts) => {
         token = await TestERC20Mintable.new('TestERC20Mintable', 'TST', 18);
         uniswapFactory = await UniswapFactory.new();
         uniswapExchange = await UniswapExchange.at(
-            (await uniswapFactory.launchExchange(token.address)).logs[0].args.exchange,
+            (await uniswapFactory.testLaunchExchange(token.address)).logs[0].args.exchange,
         );
         await token.mint(initialiser1, ether('1'));
         await token.approve(uniswapExchange.address, ether('1'), { from: initialiser1 });
@@ -94,7 +94,7 @@ contract('UniswapExchange - Initialization', (accounts) => {
 contract('UniswapExchange - Trades', (accounts) => {
 
     let token: TestERC20MintableInstance;
-    let uniswapFactory: UniswapFactoryInstance;
+    let uniswapFactory: TestUniswapFactoryInstance;
     let uniswapExchange: TestUniswapExchangeInstance;
     const initialiser1 = accounts[1];
     const initialiser2 = accounts[2];
@@ -105,7 +105,7 @@ contract('UniswapExchange - Trades', (accounts) => {
         token = await TestERC20Mintable.new('TestERC20Mintable', 'TST', 18);
         uniswapFactory = await UniswapFactory.new();
         uniswapExchange = await UniswapExchange.at(
-            (await uniswapFactory.launchExchange(token.address)).logs[0].args.exchange,
+            (await uniswapFactory.testLaunchExchange(token.address)).logs[0].args.exchange,
         );
         await token.mint(initialiser1, ether('1'));
         await token.approve(uniswapExchange.address, ether('1'), { from: initialiser1 });
@@ -560,170 +560,6 @@ contract('UniswapExchange - Trades', (accounts) => {
         await expectRevert(
             uniswapExchange.divestLiquidity(500, ether('1'), ether('1'), { from: initialiser1 }),
             'Tried to divest too much.',
-        );
-    });
-
-    /**
-     * @test {UniswapExchange#ethToToken}
-     */
-    it('eth to token internal', async () => {
-        const { tx } = await uniswapExchange.testEthToToken(
-            swapper1,
-            swapper2,
-            ether('0.5').toString(),
-            ether('0.3').toString(),
-            { from: swapper1 },
-        );
-        expectEvent.inTransaction(
-            tx,
-            'EthToTokenPurchase',
-            {
-                buyer: swapper1,
-                ethIn: ether('0.5'),
-                tokensOut: await token.balanceOf(swapper2),
-            },
-        );
-    });
-
-    /**
-     * @test {UniswapExchange#ethToToken}
-     */
-    it('eth to token internal with tokens out of range reverts', async () => {
-        await expectRevert(
-            uniswapExchange.testEthToToken.call(
-                swapper1,
-                swapper2,
-                ether('0.5'),
-                ether('0.5'),
-            ),
-            'tokensOut not in range.',
-        );
-    });
-
-    /**
-     * @test {UniswapExchange#tokenToEth}
-     */
-    it('token to eth internal', async () => {
-        const tokenAmount = ether('0.5');
-        token.mint(swapper1, tokenAmount);
-        token.approve(uniswapExchange.address, tokenAmount, { from: swapper1 });
-        const tracker2 = await balance.tracker(swapper2);
-        tracker2.get();
-        expectEvent(
-            await uniswapExchange.testTokenToEth(
-                swapper1,
-                swapper2,
-                tokenAmount,
-                ether('0.3'),
-            ),
-            'TokenToEthPurchase',
-            {
-                buyer: swapper1,
-                tokensIn: tokenAmount,
-                ethOut: await tracker2.delta(),
-            },
-        );
-    });
-
-    /**
-     * @test {UniswapExchange#ethToToken}
-     */
-    it('token to eth internal with ether out of range reverts', async () => {
-        const tracker2 = await balance.tracker(swapper2);
-        tracker2.get();
-        await expectRevert(
-            uniswapExchange.testTokenToEth(
-                swapper1,
-                swapper2,
-                ether('0.5'),
-                ether('0.5'),
-            ),
-            'ethOut not in range',
-        );
-    });
-
-    /**
-     * @test {UniswapExchange#tokenToTokenOut}
-     */
-    it('Token to token out internal', async () => {
-        // Initialize another exchange
-        const token2 = await TestERC20Mintable.new('TestERC20Mintable2', 'TST2', 18);
-        const uniswapExchange2 = await UniswapExchange.at(
-            (await uniswapFactory.launchExchange(token2.address)).logs[0].args.exchange,
-        );
-        await token2.mint(initialiser2, ether('1'));
-        await token2.approve(uniswapExchange2.address, ether('1'), { from: initialiser2 });
-        await uniswapExchange2.initializeExchange.sendTransaction(
-            ether('1'),
-            { from: initialiser2, value: (ether('1')).toString() },
-        );
-        // Swap tokens
-        const timeout = Math.floor((new Date().getTime()) / 1000) + 3600;
-        const tokenAmount = ether('0.5');
-        const minEth = ether('0.2');
-        token.mint(swapper1, tokenAmount);
-        token.approve(uniswapExchange.address, tokenAmount, { from: swapper1 });
-        const swap = await uniswapExchange.testTokenToTokenOut(
-            token2.address,
-            swapper1,
-            swapper2,
-            tokenAmount,
-            minEth,
-        );
-        expectEvent(
-            swap,
-            'TokenToEthPurchase',
-            {
-                buyer: swapper1,
-                tokensIn: tokenAmount,
-            },
-        );
-        expectEvent(
-            swap,
-            'EthToTokenPurchase',
-            {
-                buyer: uniswapExchange.address,
-                tokensOut: await token2.balanceOf(swapper2),
-            },
-        );
-    });
-
-    /**
-     * @test {UniswapExchange#tokenToTokenOut}
-     */
-    it('Token to token out internal with invalid purchased token address reverts', async () => {
-        const timeout = Math.floor((new Date().getTime()) / 1000) + 3600;
-        const tokenAmount = ether('0.5');
-        const minEth = ether('0.2');
-        await expectRevert(
-            uniswapExchange.testTokenToTokenOut(
-                constants.ZERO_ADDRESS,
-                swapper1,
-                swapper2,
-                tokenAmount,
-                minEth,
-            ),
-            'Invalid purchased token address.',
-        );
-    });
-
-    /**
-     * @test {UniswapExchange#tokenToTokenOut}
-     */
-    it('Token to token out internal with invalid exchange address reverts', async () => {
-        const token2 = await TestERC20Mintable.new('TestERC20Mintable2', 'TST2', 18);
-        const timeout = Math.floor((new Date().getTime()) / 1000) + 3600;
-        const tokenAmount = ether('0.5');
-        const minEth = ether('0.2');
-        await expectRevert(
-            uniswapExchange.testTokenToTokenOut(
-                token2.address,
-                swapper1,
-                swapper2,
-                tokenAmount,
-                minEth,
-            ),
-            'Invalid exchange address.',
         );
     });
