+++ truffle-config.js
@@ -85,7 +85,7 @@ module.exports = {
             version: '0.5.10', // Fetch exact version from solc-bin (default: truffle's version)
             settings: { // See the solidity docs for advice about optimization and evmVersion
                 optimizer: {
-                    enabled: false,
+                    enabled: true,
                     runs: 200,
                 },
                 evmVersion: 'constantinople',
